pyproject.toml
[tool.poetry]
name = "mlops-project"
version = "0.1.0"
description = ""
authors = ["Diana Anisimova <diana.anisimova.01@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
numpy = "^1.26.0"
pandas = "^2.1.1"
sklearn = "^0.0.post9"
scikit-learn = "^1.3.1"
catboost = "^1.2.2"
hydra-core = "1.3"
fire = "^0.5.0"
mlflow = "^2.8.1"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.4.0"

[tool.black]
line-length = 90
target-version = ["py39"]

[tool.isort]
src_paths = ["somepytools"]
profile = "black"
line_length = 90
lines_after_imports = 2

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
